feat: Add order deletion functionality

+++ src/public/js/book.admin.js
@@ -123,7 +123,12 @@ class BookFormModal {
                 <div>${book.SoLuongTon}</div>`
         } catch (error) {
             console.log(error)
-            showToast(error.message, 'danger')
+            Swal.close()
+            Swal.fire({
+                icon: 'error',
+                title: 'Chọn ảnh bìa thất bại!',
+                text: error.message,
+            })
         }
 
         this.imagePreview.src = this.imageCurrent
@@ -159,6 +164,15 @@ class BookFormModal {
                 formData.append('HinhAnh', this.imageInput.files[0])
             }
 
+            Swal.fire({
+                title: 'Đang xử lý...',
+                text: 'Vui lòng chờ trong giây lát!',
+                allowOutsideClick: false,
+                didOpen: () => {
+                    Swal.showLoading()
+                },
+            })
+
             const res = await fetch('/api/book', {
                 method: 'POST',
                 body: formData,
@@ -174,13 +188,27 @@ class BookFormModal {
                 throw new Error(errorMessage)
             }
 
-            showToast('Đã thêm sách', 'success')
+            Swal.close()
+
+            Swal.fire({
+                title: 'Thêm sách thành công!',
+                icon: 'success',
+                draggable: true,
+            })
+
             this.addModal.querySelector('.btn-close').click()
 
             this.bookTableInstance.updateView(1)
         } catch (error) {
             console.log(error)
-            showToast(error.message, 'danger')
+
+            Swal.close()
+
+            Swal.fire({
+                icon: 'error',
+                title: 'Thêm sách thất bại!',
+                text: error.message,
+            })
         }
     }
 
@@ -202,6 +230,15 @@ class BookFormModal {
                 formData.append('HinhAnh', this.imageInput.files[0])
             }
 
+            Swal.fire({
+                title: 'Đang xử lý...',
+                text: 'Vui lòng chờ trong giây lát!',
+                allowOutsideClick: false,
+                didOpen: () => {
+                    Swal.showLoading()
+                },
+            })
+
             const res = await fetch('/api/book/' + id, {
                 method: 'PUT',
                 body: formData,
@@ -217,7 +254,14 @@ class BookFormModal {
                 throw new Error(errorMessage)
             }
 
-            showToast('Đã cập nhật sách', 'success')
+            Swal.close()
+
+            Swal.fire({
+                title: 'Chỉnh sửa sách thành công!',
+                icon: 'success',
+                draggable: true,
+            })
+
             this.addModal.querySelector('.btn-close').click()
 
             const urlParams = new URLSearchParams(window.location.search)
@@ -226,7 +270,14 @@ class BookFormModal {
             this.bookTableInstance.updateView(page)
         } catch (error) {
             console.log(error)
-            showToast(error.message, 'danger')
+
+            Swal.close()
+
+            Swal.fire({
+                icon: 'error',
+                title: 'Chỉnh sửa sách thất bại!',
+                text: error.message,
+            })
         }
     }
 
@@ -531,17 +582,44 @@ class BookTable {
                 }
             }
         } catch (error) {
-            showToast(error.message, 'danger')
+            Swal.close()
+            Swal.fire({
+                icon: 'error',
+                title: 'Cập nhật giao diện thất bại!',
+                text: error.message,
+            })
         }
     }
 
     async deleteEntity(btnDelete) {
+        const result = await Swal.fire({
+            title: 'Bạn có chắc muốn tiếp tục xóa?',
+            text: 'Bạn sẽ không hoàn tác được sau khi xóa!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Xác nhận',
+            cancelButtonText: 'Hủy bỏ',
+        })
+
+        if (!result.isConfirmed) return
+
         const rowElement = btnDelete.closest('tr')
         const entityId = rowElement.dataset.id
 
         if (!entityId) return
 
         try {
+            Swal.fire({
+                title: 'Đang xử lý...',
+                text: 'Vui lòng chờ trong giây lát!',
+                allowOutsideClick: false,
+                didOpen: () => {
+                    Swal.showLoading()
+                },
+            })
+
             const res = await fetch(`${this.config.apiBaseUrl}/${entityId}`, {
                 method: 'DELETE',
             })
@@ -562,10 +640,23 @@ class BookTable {
             if (dataAttributeElement.dataset.totalItemPerPage < 2)
                 targetPage -= 1
 
+            Swal.close()
+
+            Swal.fire({
+                title: 'Xóa sách thành công!',
+                icon: 'success',
+                draggable: true,
+            })
+
             this.updateView(targetPage)
-            showToast(`Đã xóa ${this.config.entityName}`, 'success')
         } catch (error) {
-            showToast(error.message, 'danger')
+            Swal.close()
+
+            Swal.fire({
+                icon: 'error',
+                title: 'Xóa sách thất bại!',
+                text: error.message,
+            })
         }
     }
 
@@ -677,7 +768,12 @@ class BookTable {
             window.URL.revokeObjectURL(url)
             document.body.removeChild(a)
         } catch (error) {
-            showToast(error, 'danger')
+            Swal.close()
+            Swal.fire({
+                icon: 'error',
+                title: 'Xuất file thất bại!',
+                text: error.message,
+            })
         }
     }
 }
